Fix compilation issues on LoongArch and RISC-V
fscan uses a forked tls module, which fails to build on LoongArch and
RISC-V (despite the RISC-V changes on that fork). The forked tls
module was also used in the grdp module.

Remove dependency to the forked tls module in both the main fscan
source and in the grdp module. The latter was forked yet again under
my name until the upstream adopts this change.

Signed-off-by: Neko205 <neko200553@gmail.com>